evaluate attack on single source all targets

+++ evaluate_attack.py
@@ -421,7 +421,7 @@ def attack(sess):
 			np.save('%s/points_%d.npy' % (opt.save_dir, iter_), Vpred[0][0])
 
 			for image_index in range(adv_img.shape[0]):
-				imageio.imwrite('%s/adv_image_%d_%d.png' % (opt.save_dir, image_index, iter_), adv_img[image_index]*255)
+				imageio.imwrite('%s/adv_image_%d_%d.png' % (opt.save_dir, image_index, iter_), (adv_img[image_index]* 255).astype(np.uint8))
 
 with tf.Session(config=tfConfig) as sess:
 	util.restoreModel(opt, sess, saver)
